backend/requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
psycopg2-binary
python-multipart
PyJWT
passlib
bcrypt
PyMuPDF
python-docx
python-dotenv
slowapi
google-genai
faiss-cpu
numpy
email-validator
rank_bm25==0.2.2